review comments + edge cases
